kgsl: Add support for using compatible string "qcom,kgsl"
The drm driver tries to probe kgsl device as the compatible string
"qcom,kgsl-3d0" is part of its devicetree match table. This is not
desirable in a situation where both drm and kgsl drivers are present.
When drm driver is required for display and kgsl for GPU, a new
compatible string "qcom,kgsl" is used which is not part of drm driver's
devicetree match table. Hence add support for "qcom,kgsl" to prevent
drm driver probing kgsl device.

Change-Id: Ica8cc4e5cc75025cbdb804277171f1a4b160bcad
Signed-off-by: Puranam V G Tejaswi <quic_pvgtejas@quicinc.com>